+++ libcnb-data/src/build.rs
@@ -1,8 +1,6 @@
-use crate::bom;
 use serde::Serialize;
 
 #[derive(Serialize, Debug)]
 pub struct Build {
-    pub bom: bom::Bom,
     pub unmet: Vec<String>,
 }