+++ libcnb-data/src/launch.rs
@@ -1,13 +1,11 @@
-use crate::bom::Bom;
 use crate::newtypes::libcnb_newtype;
-use serde::{Deserialize, Serialize};
+use serde::{Deserialize, Serialize, Serializer};
+use std::path::PathBuf;
 
 /// Data Structure for the launch.toml file.
 #[derive(Deserialize, Serialize, Clone, Debug, Default)]
 #[serde(deny_unknown_fields)]
 pub struct Launch {
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub bom: Bom,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub labels: Vec<Label>,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -59,12 +57,6 @@ impl LaunchBuilder {
             self.process(process);
         }
 
-        self
-    }
-
-    /// Adds a BOM to the launch configuration.
-    pub fn bom<B: Into<Bom>>(&mut self, bom: B) -> &mut Self {
-        self.launch.bom = bom.into();
         self
     }
 
@@ -123,6 +115,54 @@ pub struct Process {
     pub direct: bool,
     #[serde(default, skip_serializing_if = "std::ops::Not::not")]
     pub default: bool,
+    #[serde(
+        rename = "working-directory",
+        default,
+        skip_serializing_if = "WorkingDirectory::is_app"
+    )]
+    pub working_directory: WorkingDirectory,
+}
+
+#[derive(Deserialize, Debug, Clone, PartialEq, Eq)]
+#[serde(untagged)]
+pub enum WorkingDirectory {
+    // There is no explicitly defined value in the CNB spec that denotes the app directory. Since
+    // we cannot enforce skipping serialization (which indicates the app directory) from this type
+    // directly, we serialize this a ".". The CNB spec says that any relative path is treated
+    // relative to the app directory, so "." will be the app directory itself. However, types that
+    // contain this type (i.e. Process), should always add
+    // `#[serde(skip_serializing_if = "WorkingDirectory::is_app")]` to a field of this type.
+    App,
+    Directory(PathBuf),
+}
+
+impl WorkingDirectory {
+    #[must_use]
+    pub fn is_app(&self) -> bool {
+        matches!(self, Self::App)
+    }
+}
+
+// Custom Serialize implementation since we want to always serialize as a string. Serde's untagged
+// enum representation does not work here since App would serialize as null, but we want a default
+// string value. #[serde(rename = ".")] doesnt work here. There are more generic solutions that can
+// be found on the web, but they're much more heavyweight than this simple Serialize implementation.
+impl Serialize for WorkingDirectory {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        match self {
+            WorkingDirectory::App => serializer.serialize_str("."),
+            WorkingDirectory::Directory(path) => path.serialize(serializer),
+        }
+    }
+}
+
+impl Default for WorkingDirectory {
+    fn default() -> Self {
+        WorkingDirectory::App
+    }
 }
 
 pub struct ProcessBuilder {
@@ -147,6 +187,7 @@ impl ProcessBuilder {
     /// * No arguments to the process
     /// * `direct` is `false`
     /// * `default` is `false`
+    /// * `working_directory` will be `WorkingDirectory::App`.
     pub fn new(r#type: ProcessType, command: impl Into<String>) -> Self {
         Self {
             process: Process {
@@ -155,6 +196,7 @@ impl ProcessBuilder {
                 args: Vec::new(),
                 direct: false,
                 default: false,
+                working_directory: WorkingDirectory::App,
             },
         }
     }
@@ -210,6 +252,12 @@ impl ProcessBuilder {
     /// default during the export phase.
     pub fn default(&mut self, value: bool) -> &mut Self {
         self.process.default = value;
+        self
+    }
+
+    /// Set the working directory for the process.
+    pub fn working_directory(&mut self, value: WorkingDirectory) -> &mut Self {
+        self.process.working_directory = value;
         self
     }
 
@@ -276,6 +324,7 @@ libcnb_newtype!(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use serde_test::{assert_ser_tokens, Token};
 
     #[test]
     fn launch_builder_add_processes() {
@@ -337,7 +386,8 @@ command = "foo"
                 command: String::from("foo"),
                 args: vec![],
                 direct: false,
-                default: false
+                default: false,
+                working_directory: WorkingDirectory::App
             })
         );
     }
@@ -359,6 +409,7 @@ command = "foo"
     fn process_with_some_default_values_serialization() {
         let process = ProcessBuilder::new(process_type!("web"), "foo")
             .default(true)
+            .working_directory(WorkingDirectory::Directory(PathBuf::from("dist")))
             .build();
 
         let string = toml::to_string(&process).unwrap();
@@ -367,6 +418,7 @@ command = "foo"
             r#"type = "web"
 command = "foo"
 default = true
+working-directory = "dist"
 "#
         );
     }
@@ -382,7 +434,8 @@ default = true
                 command: String::from("java"),
                 args: vec![],
                 direct: false,
-                default: false
+                default: false,
+                working_directory: WorkingDirectory::App
             }
         );
 
@@ -395,7 +448,8 @@ default = true
                 command: String::from("java"),
                 args: vec![],
                 direct: false,
-                default: true
+                default: true,
+                working_directory: WorkingDirectory::App
             }
         );
 
@@ -408,7 +462,22 @@ default = true
                 command: String::from("java"),
                 args: vec![],
                 direct: true,
-                default: true
+                default: true,
+                working_directory: WorkingDirectory::App
+            }
+        );
+
+        process_builder.working_directory(WorkingDirectory::Directory(PathBuf::from("dist")));
+
+        assert_eq!(
+            process_builder.build(),
+            Process {
+                r#type: process_type!("web"),
+                command: String::from("java"),
+                args: vec![],
+                direct: true,
+                default: true,
+                working_directory: WorkingDirectory::Directory(PathBuf::from("dist"))
             }
         );
     }
@@ -431,8 +500,27 @@ default = true
                     String::from("bar"),
                 ],
                 direct: false,
-                default: false
+                default: false,
+                working_directory: WorkingDirectory::App
             }
+        );
+    }
+
+    #[test]
+    fn process_working_directory_serialization() {
+        assert_ser_tokens(&WorkingDirectory::App, &[Token::BorrowedStr(".")]);
+
+        assert_ser_tokens(
+            &WorkingDirectory::Directory(PathBuf::from("/")),
+            &[Token::BorrowedStr("/")],
+        );
+        assert_ser_tokens(
+            &WorkingDirectory::Directory(PathBuf::from("/foo/bar")),
+            &[Token::BorrowedStr("/foo/bar")],
+        );
+        assert_ser_tokens(
+            &WorkingDirectory::Directory(PathBuf::from("relative/foo/bar")),
+            &[Token::BorrowedStr("relative/foo/bar")],
         );
     }
 }