+++ libcnb-data/src/lib.rs
@@ -4,7 +4,6 @@
 // This lint is too noisy and enforces a style that reduces readability in many cases.
 #![allow(clippy::module_name_repetitions)]
 
-pub mod bom;
 pub mod build;
 pub mod build_plan;
 pub mod buildpack;
@@ -13,6 +12,7 @@ pub mod exec_d;
 pub mod launch;
 pub mod layer;
 pub mod layer_content_metadata;
+pub mod sbom;
 pub mod store;
 
 mod newtypes;